Add tests to CLI
