+++ Tryouts/Prototypes/Shell/Node-launcher/Demo/package.json
@@ -5,18 +5,7 @@
   "main": "example.js",
   "module": "example.js",
   "type": "module",
-  "scripts": {
-    "clean": "rimraf output",
-    "build": "npm run clean && rollup -c"
-  },
   "dependencies": {
     "@morgan-stanley/composeui-node-launcher": "*"
-  },
-  "devDependencies": {
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "rimraf": "4.1.2",
-    "rollup": "^2.76.0",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4"
   }
 }
