--- Tryouts/Prototypes/Shell/Node-launcher/Demo/rollup.config.js
@@ -1,13 +1,0 @@
-﻿import { nodeResolve } from '@rollup/plugin-node-resolve';
-
-export default {
-    input: 'example.js',
-    output: {
-        file: 'output/bundle.js',
-        format: 'es'
-    },
-    plugins: [
-        nodeResolve()
-        
-    ]
-};
