+++ Tryouts/Prototypes/Shell/Node-launcher/Lib/package.json
@@ -7,22 +7,19 @@
   "type": "module",
   "scripts": {
     "clean": "rimraf output",
-    "build": "npm run clean && tsc && rollup -c",
+    "build": "npm run clean && tsc",
     "test": "jest"
   },
   "bin": {
-    "composeui": "./cli/cli.js"
+    "composeui": "./output/cli/cli.js"
   },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-typescript": "11.0.0",
-    "@types/node": "^18.11.18",
     "@types/jest": "29.4.0",
+    "@types/node": "^18.11.18",
     "jest": "29.4.3",
+    "rimraf": "4.1.2",
     "ts-jest": "29.0.5",
     "ts-node": "10.9.1",
-    "rimraf": "4.1.2",
-    "rollup": "^2.76.0",
     "tslib": "^2.4.0",
     "typescript": "^4.7.4"
   }
