+++ Tryouts/Prototypes/Shell/Node-launcher/Lib/src/cli/cli.ts
@@ -1,0 +1,8 @@
+﻿#!/usr/bin/env node
+
+"use strict";
+
+import { executeScriptFile } from './executeScrtiptFile.js';
+
+let fileName = process.argv.slice(2)[0];
+executeScriptFile(fileName);
