+++ Tryouts/Prototypes/Shell/Node-launcher/Lib/src/cli/executeScriptFile.spec.ts
@@ -1,0 +1,11 @@
+import { executeScriptFile } from './executeScrtiptFile';
+
+describe('CLI', () => {
+  let testFileName: string;
+
+  test('executeScriptFile() - No filename specified', () => {
+    expect(() => {
+      executeScriptFile(testFileName);
+    }).toThrow("Specify filename.");
+  });
+});
