+++ Tryouts/Prototypes/Shell/Node-launcher/Lib/src/index.ts
@@ -1,3 +1,5 @@
 ﻿export { BrowserWindow } from './BrowserWindow.js';
 export { Launcher } from './launcher.js';
 export { WindowConfig } from './WindowConfig.js';
+export * from './cli/cli.js';
+export { executeScriptFile } from './cli/executeScrtiptFile.js';
